requirements.txt
Pillow
airium